+++ .eslintrc.cjs
@@ -32,6 +32,7 @@ module.exports = {
     node: true,
   },
   rules: {
+    '@typescript-eslint/no-unused-vars': 'warn',
     '@typescript-eslint/explicit-function-return-type': 'warn',
     '@typescript-eslint/explicit-module-boundary-types': 'warn',
     '@typescript-eslint/no-explicit-any': 'warn',
